feat(categories): refactor service imports and add ValidateCategoryOwnershipService
