+++ src/modules/categories/categories.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { CategoriesService } from './categories.service';
+import { CategoriesService } from './services/categories.service';
 import { ActiveUserId } from 'src/shared/decorators/active-user';
 
 @Controller('categories')
