+++ src/modules/categories/categories.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
-import { CategoriesService } from './categories.service';
+import { CategoriesService } from './services/categories.service';
 import { CategoriesController } from './categories.controller';
+import { ValidateCategoryOwnershipService } from './services/validate-category-ownership.service';
 
 @Module({
   controllers: [CategoriesController],
-  providers: [CategoriesService],
+  providers: [ValidateCategoryOwnershipService, CategoriesService],
+  exports: [ValidateCategoryOwnershipService],
 })
 export class CategoriesModule {}
